Move clr module to separate file
